Weird test results explained.
When the initial staking period starts, there's no way that someone stakes at the same time the contract started, so no one can 100% of the rewards.

My results shows that User1, after staking at the very beginning until the end of the seven days; receives 100% of the allocated reward when he's the only stakeholder.